requirements.txt
flask
opencv-python
scikit-image
scikit-learn
pypotrace
boto3
cuid
gevent
sentry-sdk
Pillow
requests
pytest
faiss-cpu